infrasyntax/parser/requirements.txt
elasticsearch==8.9.0
sentence-transformers==2.7.0
torch==2.1.0
tqdm==4.66.1
python-dotenv==1.0.0
